updated code

+++ adminlte/simplecrud/delete.php
@@ -1,21 +1,31 @@
-<?php include '../db.php';
-// $servername = "localhost";
-// $username = "root";
-// $password = "admin123";
-// $dbname = "guest";
-
-// Create connection
-// $conn = new mysqli($servername, $username, $password,$dbname);
-
-// Check connection
-// if ($conn->connect_error) {
-//   die("Connection failed: " . $conn->connect_error);
-// }
-
-
-$id = $_GET['id'];
+<?php
+session_start();
+include '../db.php';
  
-mysqli_query($conn, "DELETE FROM users WHERE user_id=$id");
+if (!isset($_GET['id'])) {
+    die("User id missing");
+}
+ 
+$id = intval($_GET['id']); 
+ 
+
+$loggedInUserId = $_SESSION['user_id'] ?? null;
+ 
+
+mysqli_query($conn, "DELETE FROM users WHERE user_id = $id");
+ 
+
+if ($loggedInUserId == $id) {
+ 
+  
+    session_unset();
+    session_destroy();
+ 
+    header("Location: ../login.php");
+    exit();
+}
+ 
 
 header("Location: listing.php");
+exit();
 ?>
